Upload  SVG icons for use
We are switching the JS map widgets to use SVG assets in order to be
more scalable. This ensures they're actually available.

Right now, this basically just copies all the SVGs listed in the icon
input JSON file to an output directory and uploads its contents to S3,
but it could be changed to optimise them for web use in the future.

Buddy Jonty.